Moved render out from mithril interop layer
