+++ packages/commonwealth/client/scripts/helpers/index.ts
@@ -1,5 +1,4 @@
 import type { ClassComponent } from 'mithrilInterop';
-import { render } from 'mithrilInterop';
 import BigNumber from 'bignumber.js';
 import { ChainBase, ChainNetwork } from 'common-common/src/types';
 import moment from 'moment';
@@ -8,6 +7,7 @@ import type { Coin } from 'adapters/currency';
 import Account from '../models/Account';
 import IChainAdapter from '../models/IChainAdapter';
 import { ThreadStage } from '../models/types';
+import { render } from 'helpers/DEPRECATED_ReactRender';
 
 export async function sleep(msec) {
   return new Promise((resolve) => setTimeout(resolve, msec));
