+++ packages/commonwealth/client/scripts/mithrilInterop/index.ts
@@ -9,73 +9,16 @@
  */
 
 import type { FunctionComponent, ReactNode } from 'react';
-import { createElement, Component as ReactComponent } from 'react';
+import { Component as ReactComponent } from 'react';
 import type {
   NavigateFunction,
   Location,
   NavigateOptions,
 } from 'react-router-dom';
-import { createRoot } from 'react-dom/client';
 import { getScopePrefix } from 'navigation/helpers';
 
 // corresponds to Mithril's "Children" type -- RARELY USED
 export type Children = ReactNode | ReactNode[];
-
-// corresponds to Mithril's `m()` call
-export const render = (
-  tag: string | React.ComponentType,
-  attrs: any = {},
-  ...children: any[]
-) => {
-  let className = attrs?.className;
-
-  // check if selector uses classes (.) and convert to props, as
-  // Mithril supports e.g. `.User.etc` tags while React does not
-  if (typeof tag === 'string' && tag.includes('.')) {
-    const [selector, ...classes] = tag.split('.');
-
-    if (className) {
-      className = `${className} ${classes.join(' ')}`;
-    } else {
-      className = classes.join(' ');
-    }
-
-    // replace pure class selector (e.g. .User) with div (e.g. div.User)
-    tag = selector || 'div';
-  }
-
-  // handle children without attrs => corresponds to `m(tag, children)`
-  if (Array.isArray(attrs) || typeof attrs !== 'object') {
-    children = attrs;
-    attrs = { className };
-  } else {
-    attrs = { ...attrs };
-  }
-
-  // ensure vnode.className exists
-  attrs.className = className;
-
-  // react forbids <img> and <br> tags to have children
-  if (tag === 'img' || tag == 'br') {
-    return createElement(tag, attrs);
-  }
-
-  // ensure vnode.children exists as mithril expects
-  attrs.children = children;
-
-  return createElement(tag, attrs, ...children);
-};
-
-// corresponds to Mithril's `m.trust()` call, which is used to render inner HTML directly
-render.trust = (html: string, wrapper?: string) => {
-  if (wrapper) {
-    return createElement(wrapper, {
-      dangerouslySetInnerHTML: { __html: html },
-    });
-  } else {
-    return createElement('div', { dangerouslySetInnerHTML: { __html: html } });
-  }
-};
 
 // Mithril component type, maps to FC
 export type Component<Props = unknown> = FunctionComponent<Props>;
@@ -207,14 +150,4 @@ export function redraw(sync = false, component?: any) {
   } else {
     // console.trace('no-op redraw called!');
   }
-}
-
-// m.mount() shim
-export function rootMount(element: Element, component?: any | null) {
-  return createRoot(element).render(component);
-}
-
-// m.render() shim
-export function rootRender(el: Element, vnodes: Children) {
-  return rootMount(el, render('div', {}, vnodes));
 }
