+++ packages/commonwealth/client/scripts/models/ChainInfo.ts
@@ -1,4 +1,4 @@
-import { render } from 'mithrilInterop';
+import { render } from 'helpers/DEPRECATED_ReactRender';
 import $ from 'jquery';
 import type { RegisteredTypes } from '@polkadot/types/types';
 import app from 'state';
