+++ packages/commonwealth/client/scripts/models/MinimumProfile.ts
@@ -1,4 +1,4 @@
-import { render } from 'mithrilInterop';
+import { render } from 'helpers/DEPRECATED_ReactRender';
 import jdenticon from 'jdenticon';
 
 import {
